Compile saved configuration parser contracts
