+++ tests/migration-ledger.md
@@ -8,7 +8,7 @@ equivalent protection.
 | Rule family | Defect protected against | Replacement layer | Mutation case | Status |
 |---|---|---|---|---|
 | `check_firmware_ha_bindings.py` Home Assistant read, callback and subscription shapes | Lost updates, unsafe callback lifetime, duplicated reads, stale reconnect work | Host C++ `HaReadCoordinator` executable tests | Duplicate-read fan-out mutation plus disconnect, low heap, re-entry, reconnect, cancel and stale-generation tests | Stage 2 overlap verified; old rule retained until the migration is complete |
-| `check_firmware_parser.py` extracted parser source | Saved configurations becoming unreadable or normalising differently | Host C++ tests compiling the real parser headers with shared JSON fixtures | Legacy input rejected or canonical bytes changed | Planned in stage 3; old rule retained |
+| `check_firmware_parser.py` extracted parser source | Saved configurations becoming unreadable or normalising differently | CTest `saved_config_parser_test` compiles the production parser header with the same JSON fixtures used by the real web parser tests | Compact marker mutation makes legacy or canonical configurations unreadable | Stage 3 overlap verified; JavaScript firmware-parser imitation retired and old compiled check retained |
 | `check_firmware_card_runtime.py` runtime source patterns | Card actions or rendering paths silently disappearing | Focused host C++ domain tests | Remove the protected action or state transition | Awaiting executable domain migration (#1007); old rule retained |
 | `check_firmware_modals.py` modal source patterns | Broken navigation, modal lifecycle, controls or generated wiring | Host C++ state tests and browser interaction tests | Remove lifecycle transition or generated binding | Awaiting executable domain migration (#1007); old rule retained |
 | `check_firmware_display_tokens.py` display token scanning | Unknown or misplaced display substitutions reaching firmware | Structural schema/token validator | Add an invalid token at a guarded location | Genuine structural rule; retain |
